# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aelu API
  version: 1.0.0
  description: >
    REST API for the Aelu learning system. Supports adaptive drill sessions
    (via WebSocket), spaced-repetition progress tracking, graded reading, media
    exposure, extensive listening, vocabulary encounter logging, xAPI/Caliper
    learning-record export, Common Cartridge export, GDPR data portability and
    erasure, JWT-based mobile auth, offline sync, Stripe payments, and push
    notifications.
servers:
  - url: https://aelu.app
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/export/common-cartridge:
    get:
      tags: [Export]
      summary: Export vocabulary as IMS Common Cartridge ZIP
      operationId: exportCommonCartridge
      parameters:
        - name: level
          in: query
          schema:
            type: integer
            default: 1
          description: HSK level to export
      responses:
        "200":
          description: Common Cartridge ZIP file
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string

  # ── Reading ─────────────────────────────────────────────────────────
...
  /api/xapi/statements:
    get:
      tags: [xAPI]
      summary: Retrieve xAPI statements for the authenticated user
      operationId: getXapiStatements
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only return statements after this timestamp
        - name: until
          in: query
          schema:
            type: string
            format: date-time
          description: Only return statements before this timestamp
      responses:
        "200":
          description: xAPI statements
          content:
            application/json:
              schema:
                type: object
                properties:
                  statements:
                    type: array
                    items:
                      type: object
                      description: xAPI statement per spec

  # ── Caliper ─────────────────────────────────────────────────────────
  /api/caliper/events:
    get:
      tags: [Caliper]
      summary: Retrieve IMS Caliper 1.2 events for the authenticated user
      operationId: getCaliperEvents
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only return events after this timestamp
      responses:
        "200":
          description: Caliper events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: object
                      description: Caliper event envelope

  # ── GDPR ────────────────────────────────────────────────────────────
...
  /api/onboarding/level:
    post:
      tags: [Onboarding]
      summary: Set initial HSK level during onboarding
      operationId: setOnboardingLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
  /api/onboarding/goal:
    post:
      tags: [Onboarding]
      summary: Set learning goal during onboarding
      operationId: setOnboardingGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goal:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
...
  /api/onboarding/placement/submit:
    post:
      tags: [Onboarding]
      summary: Submit placement test answers
      operationId: submitPlacement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object

  # ── Grammar ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie set on login
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/token
  schemas:
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok
...